move socket.py into device.py
